% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_homa.R
\name{cs_homa}
\alias{cs_homa}
\title{HOMA-IR}
\usage{
cs_homa(insulin, insulin_unit, glucose)
}
\arguments{
\item{insulin}{insulin value}

\item{insulin_unit}{insulin unit (pg/ml or mU/l)}

\item{glucose}{glucose value}
}
\description{
Calulate HOMA-IR from Insulin and Glucose
}
\details{
This function calculates HOMA-IRfrom Insulin and Glucose.
}
\examples{

USArrests \%>\%

USArrests \%>\%

}
