% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_to_na.R
\name{extreme_to_na}
\alias{extreme_to_na}
\title{Convert Extreme Outliers to NA}
\usage{
extreme_to_na(col_val)
}
\arguments{
\item{col_val}{The column or values to be checked}
}
\value{
Values with extreme outliers converted to NA_real_
}
\description{
Convert extreme outliers to NA
}
\details{
This function uses the boxplot method to convert extreme outliers to N. Extreme outliers
are values above Q3 + 3xIQR or below Q1 - 3xIQR.
}
\examples{

# Vector
col_val <- c(1, 2, 3, 5, 19)

extreme_to_na(col_val)

# Data frame

library(dplyr)
dat <- tibble(
  a = c(1, 2, 3, 5, 19)
  )

dat \%>\%
  mutate(new_a = extreme_to_na(a))

}
