% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bxjt.R
\name{geom_bxjt}
\alias{geom_bxjt}
\title{Boxplot + Jitter + Mean}
\usage{
geom_bxjt(..., fill = "#00A2CEFF", color = NA, size = 3)
}
\arguments{
\item{fill}{fill for geom_jitter (default = "#00A2CEFF")}

\item{color}{color for geom_jitter}

\item{size}{size of points (default = 3)}
}
\description{
Boxplot with jitter and mean
}
\details{
This function combines geom_boxplot, geom_jitter, and stat_summary("mean") for ggplot.
}
\examples{

mtcars \%>\%
ggplot(aes(as.factor(cyl), mpg)) +
geom_bxjt()

mtcars \%>\%
ggplot(aes(as.factor(cyl), mpg)) +
geom_bxjt(fill = "green", color = "white")

}
