% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_smcor.R
\name{geom_smcor}
\alias{geom_smcor}
\title{Point + Smooth + Cor}
\usage{
geom_smcor(
  ...,
  pt_fill = "#00A2CEFF",
  pt_color = "black",
  pt_size = 2,
  ln_color = "black"
)
}
\arguments{
\item{pt_fill}{fill for geom_jitter (default = "#00A2CEFF")}

\item{pt_color}{color for geom_jitter}

\item{pt_size}{size of points (default = 3)}

\item{ln_color}{color for geom_smooth}
}
\description{
Scatterplot with trendline and correlation
}
\details{
This function combines geom_point, geom_smooth, and stat_cor for ggplot.
}
\examples{

USArrests \%>\%
  ggplot(aes(Rape, Murder)) +
  geom_smcor()

USArrests \%>\%
  ggplot(aes(Rape, Murder)) +
  geom_smcor(pt_fill = "red", pt_color = "yellow", pt_size = 2, ln_color = "orange")

}
