% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_html_to_outputs.R
\name{knit_html_to_outputs}
\alias{knit_html_to_outputs}
\title{Knit html to outputs subfolder}
\usage{
knit_html_to_outputs(input, ...)
}
\arguments{
\item{input}{This is the file you want to knit. This is just left blank}
}
\description{
Knit to outputs subfolder
}
\details{
This function allows you to make your output html file go to an outputs subfolder when you click the \code{knit} button or use \code{Ctrl + Shift + K}.
You have to add this function to your YAML header after the \verb{knit:} option. It creates and html file.
}
\examples{

knit: crsp::knit_html_to_outputs

}
