% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_to.R
\name{knit_to}
\alias{knit_to}
\title{Knit to custom location}
\usage{
knit_to(input, location, outputs, ...)
}
\arguments{
\item{input}{This is the file you want to knit.}

\item{location}{This is where you want to knit to.}

\item{outputs}{This is the type of outputs you want.}
}
\description{
Knit to specific location
}
\details{
This function allows you to specify where you want your output files to end up when you click the \code{knit} button or use \code{Alt + Ctrl + k}.
You have to add this function to your YAML header after the \verb{knit:} option.
}
\examples{

knit(location = here::here("outputs"), outputs = c("html_document", "md_document"))

}
