% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_to_outputs_folder.R
\name{knit_to_outputs_folder}
\alias{knit_to_outputs_folder}
\title{Knit to outputs subfolder}
\usage{
knit_to_outputs_folder(input, ...)
}
\arguments{
\item{input}{This is the file you want to knit. This is just left blank}
}
\description{
Knit to outputs subfolder
}
\details{
This function allows you to make your output files go to an outputs subfolder when you click the \code{knit} button or use \code{Ctrl + Shift + K}.
You have to add this function to your YAML header after the \verb{knit:} option. It creates and html file and an .md file.
}
\examples{

knit: crsp::knit_to_outputs_folder

}
