% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_pal_c.R
\name{rand_pal_c}
\alias{rand_pal_c}
\title{Get Random Palette from {paletteer}}
\usage{
rand_pal_c(n_colors = 3)
}
\arguments{
\item{n_colors}{The minimum number of colors in the palette.}
}
\value{
A list of the package name and the colors
}
\description{
Get Random Continuous Color Palette
}
\details{
This function pulls a random continuous palette from {paletteer} that has a minimum length. It returns the name of the palette and the hexcodes.
}
\examples{

rand_pal_c(n_colors = 6)

}
