% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_pal_d.R
\name{rand_pal_d}
\alias{rand_pal_d}
\title{Get Random Palette from {paletteer}}
\usage{
rand_pal_d(n_colors)
}
\arguments{
\item{n_colors}{The minimum number of colors in the palette.}
}
\value{
A list of the package name and the colors
}
\description{
Get Random Discrete Color Palette
}
\details{
This function pulls a random discrete palette from {paletteer} that has a minimum length. It returns the name of the palette and the hexcodes.
}
\examples{

rand_pal_d(length = 4)

}
