% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Setup Project File Structure}
\usage{
setup_project()
}
\description{
Setup project file structure.
}
\details{
This function is used to standardize the file structure within an R Project.
}
