% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bxjt.R
\name{geom_bxjt}
\alias{geom_bxjt}
\title{Boxplot + Jitter + Mean}
\usage{
geom_bxjt(..., fill = NA, color = NA)
}
\arguments{
\item{fill}{fill for geom_jitter}

\item{color}{color for geom_jitter}
}
\description{
Boxplot with jitter and mean
}
\details{
This function combines geom_boxplot, geom_jitter, and stat_summary("mean") for ggplot.
}
\examples{

mtcars \%>\%
ggplot(aes(as.factor(cyl), mpg)) +
geom_bxjt()


}
